DDPB-2589: Add a remove confirmation page to all summary pages (#1281)
* Create common 'confirm delete' template

* Add confirmation page when deleting NDR assets

* Add valuation date to assets definition

And format value number better

* Tidy up delete confirmation page HTML

* Rename definition to summary

* Move breadcrumbs and link back outside of <main>

That's how their CSS is designed to work, otherwise, the margins are confusing

* Correct CSS class in title supports

* Use secondary button rather than fake link

* Use ConfirmDeleteType form in confirmation page

It's a little cleaner than manually getting and setting GET variables

* Show different summaries for properties and other assets

* Add delete confirmation page for bank accounts

* Add delete confirmation page for NDR deputy expenses

* Use more complex summary definition

This allows us to pass an additional `format` parameter for each part of the summary, and will make playing with the order easier (since keys are now numeric).

* Add delete confirmation page for user accounts

Replace the custom page which didn't add any additional value to our shared page

* Add clarification when deleting a user with other teams

* Update user removal translations

* Add delete confirmation page for client contacts

Replace the custom page which didn't add any additional value to our shared page

* Add delete guard for report assets

* Use entered, not calculated, value of property on deletion guard

This better matches what's shown on the summary

* Add delete guard for report bank accounts

Replaces existing bespoke page which worked the same way

* Add warning if deleting account with attached payments

* Add string util function to join lists more naturally

* Use new string util function for list of transaction types

* Identify properly whether `warning` is set on delete guard

* Add delete guard for report contacts

* Pass report to deletion guard templates

This is necessary so that the details banner is shown correctly

* Add delete guard for report decisions

* Add delete guard for lay deputy expenses

* Add delete guard for report gifts

* Add delete guard for report short money in

* Add fallback title on delete guard pages

Ensure that all delete guard pages have a suitable title

* Add delete guard for report short money out

* Only define delete guard warning if necessary

* Fix typos in short transaction controllers

* Add delete guard for report money in

* Add delete guard for report money out

* Add delete guard for report money transfers

* Fix button alignment

* Add delete guard for professional deputy costs

* Add delete guard for PA deputy expenses

* Use common deletion guard for documents

It's easier to maintain just one template, and the common page is better defined and has better functionality.

* Use common deletion guard for client notes

It's easier to maintain just one template and they were basically doing the same thing

* Use simpler method of finding transaction being deleted

`array_filter` will always go over entire array, and retains keys causing lookup problems. A for loop is faster and easier to understand.

* Only show bank account details for transactions with bank accounts attached

* Update deletion tests to use guard page

* Move subject into translation files

* Reinstate transfer deletion error message

* Remove now useless summaryPage translation

* Remove now useless summaryPage translation

* Move delete page translations to separate file

* Allow label and value to be translation keys

* Move NDR labels into translation files

* Change translation key to better match entity property

* Move lay report labels into translation files

* Move non-lay report labels into translation files

* Translate property string

* Add spacing between confirm and cancel buttons

Since this is a dangerous action, the spacing ensures that users don't accidentally press the wrong button

+++ src/AppBundle/Controller/Ndr/AssetController.php
@@ -338,18 +338,48 @@ class AssetController extends AbstractController
 
     /**
      * @Route("/ndr/{ndrId}/assets/{assetId}/delete", name="ndr_asset_delete")
+     * @Template("AppBundle:Common:confirmDelete.html.twig")
      *
      * @return RedirectResponse
      */
     public function deleteAction(Request $request, $ndrId, $assetId)
     {
-        $ndr = $this->getNdrIfNotSubmitted($ndrId, self::$jmsGroups);
+        $form = $this->createForm(FormDir\ConfirmDeleteType::class);
+        $form->handleRequest($request);
 
-        if ($ndr->hasAssetWithId($assetId)) {
-            $this->getRestClient()->delete("/ndr/{$ndrId}/asset/{$assetId}");
-            $request->getSession()->getFlashBag()->add('notice', 'Asset removed');
+        if ($form->isValid()) {
+            $ndr = $this->getNdrIfNotSubmitted($ndrId, self::$jmsGroups);
+
+            if ($ndr->hasAssetWithId($assetId)) {
+                $this->getRestClient()->delete("/ndr/{$ndrId}/asset/{$assetId}");
+                $request->getSession()->getFlashBag()->add('notice', 'Asset removed');
+            }
+
+            return $this->redirect($this->generateUrl('ndr_assets', ['ndrId' => $ndrId]));
         }
 
-        return $this->redirect($this->generateUrl('ndr_assets', ['ndrId' => $ndrId]));
+        $asset = $this->getRestClient()->get("ndr/{$ndrId}/asset/{$assetId}", 'Ndr\\Asset');
+
+        if ($asset instanceof EntityDir\Ndr\AssetProperty) {
+            $summary = [
+                ['label' => 'deletePage.summary.type', 'value' => 'deletePage.summary.property', 'format' => 'translate'],
+                ['label' => 'deletePage.summary.address', 'value' => implode(', ', $asset->getAddressValidLines())],
+                ['label' => 'deletePage.summary.value', 'value' => $asset->getValue(), 'format' => 'money'],
+            ];
+        } else {
+            $summary = [
+                ['label' => 'deletePage.summary.type', 'value' => $asset->getTitle()],
+                ['label' => 'deletePage.summary.description', 'value' => $asset->getDescription()],
+                ['label' => 'deletePage.summary.value', 'value' => $asset->getValue(), 'format' => 'money'],
+                ['label' => 'deletePage.summary.valuationDate', 'value' => $asset->getValuationDate(), 'format' => 'date'],
+            ];
+        }
+
+        return [
+            'translationDomain' => 'ndr-assets',
+            'form' => $form->createView(),
+            'summary' => $summary,
+            'backLink' => $this->generateUrl('ndr_assets_summary', ['ndrId' => $ndrId]),
+        ];
     }
 }
